Rename shadow function and add documentation
